Let the public API generator own the surface it already documents
It drove `createPublicApiShoppingPublicRoutes()` and carried the other twelve
paths through from the committed document untouched, so `verify:openapi-drift`
reported all fifteen as matching their routes while twelve were compared to
nothing.

There is no second surface to add. `createPublicApiRoutes()` already mounts the
shopping routes, so it produces all fifteen on its own — driving both documented
`/v1/public/shopping/search` twice, which the duplicate guard caught the moment
it was written that way. That guard exists because the same fault in finance
reached `verify:deployment-graph-openapi-coverage` as a duplicate authority
before anything else noticed.

The drift here is metadata rather than schemas. Twelve paths carried
`tags: ["storefront"]` and `x-voyant-module: "storefront"` — retired vocabulary
frozen into the published contract from before the rename — and now derive from
the routes. `x-voyant-unit-id` leaves the per-package artifact; composition adds
it back in `buildSelectedGraphOpenApiDocuments`, and coverage reports 82
documented surface/module pairs where it reported 78.

With this, every path in every registered document is produced by its generator:
531 of 531, and the ownership baseline is empty. It was 392 of 529 this morning.

Renamed off `generate-shopping-openapi.ts`; it stopped being about shopping.
