+++ scripts/checks/openapi/generated-specs.json
@@ -119,7 +119,7 @@
       ]
     },
     {
-      "command": "pnpm --filter @voyant-travel/public-api generate:shopping-openapi",
+      "command": "pnpm --filter @voyant-travel/public-api generate:openapi",
       "files": ["packages/public-api/openapi/public-api/public-api.json"]
     },
     {
